api/requirements.txt
click==8.0.1
colorama==0.4.4
docx2pdf==0.1.7
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
mysqlclient==2.0.3
PyMySQL==1.0.2
python-docx==0.8.11
pywin32==227
SQLAlchemy==1.4.22
tqdm==4.62.2
Werkzeug==2.0.1
